Move `krel gcbmgr history` implementation into pkg
To be able to add mocked unit tests later on as well as re-using the
logic in `krel stage`, we need to have it inside of the `gcb` package.

This patch achieves this transformation without changing the overall
behavior.
